+++ src/main/java/domain/lottoGame/Lotto.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.lottoGame;
+
+import domain.Money;
 
 import java.util.Arrays;
 import java.util.List;
