+++ src/main/java/domain/lottoGame/LottoFactory.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.lottoGame;
+
+import domain.lottoGame.shuffleStrategy.ShuffleStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
