+++ src/main/java/domain/lottoGame/LottoNumber.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lottoGame;
 
 import java.util.Objects;
