+++ src/main/java/domain/lottoGame/shuffleStrategy/DefaultShuffleStrategy.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.lottoGame.shuffleStrategy;
+
+import domain.lottoGame.LottoNumber;
 
 import java.util.ArrayList;
 import java.util.Collections;
