+++ src/main/java/view/OutputView.java
@@ -1,6 +1,6 @@
 package view;
 
-import domain.*;
+import domain.lottoGame.*;
 
 import java.util.Arrays;
 import java.util.Comparator;
