+++ src/test/java/domain/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.lottoGame.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
