+++ src/test/java/domain/LottoNumbersTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.lottoGame.LottoNumber;
+import domain.lottoGame.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
