+++ src/test/java/domain/LottoTest.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.lottoGame.Lotto;
+import domain.lottoGame.LottoNumber;
+import domain.lottoGame.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
